Cleaning

+++ src/preprocessing/mfcc_extraction.py
@@ -62,29 +62,31 @@ def generate_histogram(sample, root_dir, out_dir):
     #     print(sample.target + '/' + file_name)
 
 def generate_spectrogram(sample, root_dir, out_dir):
+    file_name = sample.filename + '.png'
 
-    # start_time = int(sample.filename[-2:])+int(sample.filename[-3])*60
-    y, sr = librosa.load(os.path.join(root_dir, sample.target, sample.filename[:-4] + '.wav'), duration=8)
-    # splits = librosa.util.frame(y, frame_length=int(sr/2), hop_length=512)
-    if not os.path.exists(os.path.join(out_dir, sample.target)):
-        os.mkdir(os.path.join(out_dir, sample.target))
-    # librosa.output.write_wav(os.path.join(out_dir, sample.target, sample.filename[:-4] + '.wav'), y, sr)
-    figsize = 1.76, 1.28
-    plt.rcParams["figure.figsize"] = figsize
+    if not os.path.exists(os.path.join(out_dir, sample.target, file_name)):
 
-    S = librosa.feature.melspectrogram(y=y, n_mels=128,  hop_length=1024)
+        start_time = int(sample.filename[-2:])+int(sample.filename[-3])*60
+        y, sr = librosa.load(os.path.join(root_dir, sample.target, sample.filename[:-4] + '.wav'), duration=8, offset=start_time)
+        # splits = librosa.util.frame(y, frame_length=int(sr/2), hop_length=512)
+        if not os.path.exists(os.path.join(out_dir, sample.target)):
+            os.mkdir(os.path.join(out_dir, sample.target))
+        # librosa.output.write_wav(os.path.join(out_dir, sample.target, sample.filename[:-4] + '.wav'), y, sr)
+        figsize = 1.76, 1.28
+        plt.rcParams["figure.figsize"] = figsize
 
-    librosa.display.specshow(librosa.power_to_db(S, ref=np.max), fmax=8000)
-    dpi = 100
-    # What size does the figure need to be in inches to fit the image?
+        S = librosa.feature.melspectrogram(y=y, n_mels=128,  hop_length=1024)
 
-    plt.draw()
-    plt.subplots_adjust(left=0, right=1, top=1, bottom=0)
-    file_name = sample.filename[:-4] + '.png'
+        librosa.display.specshow(librosa.power_to_db(S, ref=np.max), fmax=8000)
+        dpi = 100
+        # What size does the figure need to be in inches to fit the image?
 
-    plt.savefig(os.path.join(out_dir, sample.target, file_name), dpi=dpi, bbox_inches=0)
-    plt.close()
-    print(sample.target + '/' + file_name)
+        plt.draw()
+        plt.subplots_adjust(left=0, right=1, top=1, bottom=0)
+
+        plt.savefig(os.path.join(out_dir, sample.target, file_name), dpi=dpi, bbox_inches=0)
+        plt.close()
+        print(sample.target + '/' + file_name)
 
     # for i in range(splits.shape[-1]):
     #     S = librosa.feature.melspectrogram(y=splits[:,i], sr=sr, n_mels=96)
